Code/qgai/requirements.txt
numpy~=2.0.2
opencv-python~=4.12.0.88
pillow~=11.1.0
joblib~=1.5.1
scikit-learn~=1.6.1
openai-whisper~=20250625
librosa~=0.11.0
ffmpeg~=7.0.1
ffmpeg-python~=0.2.0
pycryptodome~=3.23.0
websockets~=15.0.1
transformers~=4.54.0
tqdm~=4.67.1
sympy~=1.13.1
requests~=2.32.4
dotenv~=0.9.9
python-dotenv~=1.1.1